requirements.txt
python-telegram-bot==20.8
requests==2.32.4
httpx==0.26.0